+++ autoprotocol/protocol.py
@@ -795,6 +795,7 @@ class Protocol(object):
                     volumes = []
                     s = source.wells[source_counter]
                     vol = s.volume
+                    max_decimal_places = 12
                     for idx, d in enumerate(dest.wells):
                         vol_d = volume[idx]
                         while vol_d > Unit.fromstring("0:microliter"):
@@ -803,12 +804,15 @@ class Protocol(object):
                                 destinations.append(d)
                                 volumes.append(vol_d)
                                 vol -= vol_d
+                                vol.value = round(vol.value, max_decimal_places)
                                 vol_d -= vol_d
+                                vol_d.value = round(vol_d.value, max_decimal_places)
                             else:
                                 sources.append(s)
                                 destinations.append(d)
                                 volumes.append(vol)
                                 vol_d -= vol
+                                vol_d.value = round(vol_d.value, max_decimal_places)
                                 source_counter += 1
                                 if source_counter < len_source:
                                     s = source.wells[source_counter]